Reorganized stories file.
